Refactor pages manager modal
